requirements.txt
pandas
requests
nltk
vaderSentiment
scikit-learn
python-dotenv